Document and extract the structs from the downloader code
